+++ core-service/src/main/resources/db/migration/V4__orders.sql
@@ -2,7 +2,7 @@ create table orders
 (
     id          bigserial primary key,
     username    varchar(255),
-    total_price int    not null,
+    total_price numeric(8, 2) not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -14,8 +14,8 @@ create table order_items
     id                bigserial primary key,
     product_id        bigint references products (id),
     order_id          bigint references orders (id),
-    price_per_product int,
-    price             int,
+    price_per_product numeric(8, 2),
+    price             numeric(8, 2),
     quantity          int,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
